Avances en el backend: 1. Se termina la funcionalidad de los botones plus y minus, tambien al escribir en el inputText funciona los casos de letras, negativos,etc" 2. Se actualizan los precios total de un producto y el precio final total, segun la cantida de productos totales. Se actualiza tanto en el front como en el backend todo l 3. Se agrego un nuevo endPoint Put al /item/:productId/update necesario para poder modificar los productos que estan en el carrito. 4. FALTA: 1. El endpoint delete a '/shop/cart'  para eliminar un producto del carrito y actualizar la pagina luego de eliminar el producto. 	  2. El endpoint POST a  'shop/cart' para cargar el carrito para ir a pagar el carrito.

+++ models/mysql/funkoShop.sql
@@ -212,7 +212,7 @@ CREATE TABLE cart(
 	cart_id INTEGER NOT NULL,
     product_id INTEGER NOT NULL,
     quantity INTEGER, 
-    total_price_product INTEGER,
+    total_price_product decimal(10,2),
     PRIMARY KEY (cart_id, product_id),
     FOREIGN KEY (product_id) REFERENCES product(product_id) ON DELETE CASCADE
 );
